+++ src/tdme/utilities/IntBuffer.h
@@ -28,27 +28,27 @@ public:
 	/**
 	 * @return capacity
 	 */
-	inline virtual int32_t getCapacity() {
-		return Buffer::getCapacity() / sizeof(int32_t);
+	inline virtual int64_t getCapacity() {
+		return Buffer::getCapacity() / sizeof(uint32_t);
 	}
 
 	/**
 	 * @return position
 	 */
-	inline virtual int32_t getPosition() {
-		return Buffer::getPosition() / sizeof(int32_t);
+	inline virtual int64_t getPosition() {
+		return Buffer::getPosition() / sizeof(uint32_t);
 	}
 
 	/**
 	 * Get a value at given position
 	 * @param position position
 	 */
-	inline int32_t get(int32_t position) {
-		int32_t value = 0;
-		value+= ((int32_t)Buffer::get(position)) & 0xFF;
-		value+= ((int32_t)Buffer::get(position + 1) << 8) & 0xFF;
-		value+= ((int32_t)Buffer::get(position + 2) << 16) & 0xFF;
-		value+= ((int32_t)Buffer::get(position + 3) << 24) & 0xFF;
+	inline uint32_t get(int64_t position) {
+		uint32_t value = 0;
+		value+= ((uint32_t)Buffer::get(position)) & 0xFF;
+		value+= ((uint32_t)Buffer::get(position + 1) << 8) & 0xFF;
+		value+= ((uint32_t)Buffer::get(position + 2) << 16) & 0xFF;
+		value+= ((uint32_t)Buffer::get(position + 3) << 24) & 0xFF;
 		return value;
 	}
 
@@ -56,8 +56,8 @@ public:
 	 * Puts a value into buffer at its current position
 	 * @param value value
 	 */
-	inline IntBuffer* put(int32_t value) {
-		Buffer::put((const uint8_t*)&value, sizeof(int32_t));
+	inline IntBuffer* put(uint32_t value) {
+		Buffer::put((const uint8_t*)&value, sizeof(uint32_t));
 		return this;
 	}
